Fix numerical validators with precision=None
If precision is None, then I believe the sensible behaviour is not to
apply a `round`, but to instead maintain the precision of the given
value.

This change corrects the base class to apply round only in the
circumstance that precision is not None. Also true when using
`Validator.dump`.